+++ packages/conc/changelog.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Change `Race.timeout` to take a `Sem` for the fallback instead of a pure value.
+* Export all `Queue` constructors from `Polysemy.Conc.Queue`.
 
 # 0.2.0.0
 * Add `read*` constructors for `Sync`